py-requirements.txt
rich
typer[all]
pillow
ruff
polars
python-dateutil
 matplotlib
 seaborn
 pyarrow
